backend/pyproject.toml
[project]
name = "acme-dental-ai-agent"
version = "0.0.1"
description = "AI Agent for the Acme Dental Clinic"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "langchain[anthropic]>=0.3.0",
    "httpx>=0.28.1",
    "apscheduler>=3.10.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic[email]>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-httpx>=0.36.0",
]

